created client controller
